requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.25.7
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bayesian-optimization==1.4.3
bayespy==0.5.1
biopython==1.83
black==24.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
coolname==2.2.0
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.5
datasets==2.1.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
fastai==2.7.15
fastcore==1.5.33
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
future==1.0.0
gensim==4.3.2
genson==1.2.2
googleapis-common-protos==1.63.0
griffe==0.36.9
grpcio==1.63.0
guardrails-ai==0.4.3
guardrails-api-client==0.1.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
hmmlearn==0.3.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
humanize==4.8.0
hyperopt==0.2.7
idna==3.7
igraph==0.11.4
imagecodecs==2024.1.1
imageio==2.34.1
imbalanced-learn==0.12.2
imgaug==0.4.0
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jwt==1.3.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
langchain-core==0.1.50
langcodes==3.4.0
langsmith==0.1.54
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2
lightgbm==4.3.0
lightning-utilities==0.11.2
llvmlite==0.42.0
loguru==0.7.2
lxml==4.9.4
Mako==1.3.3
marisa-trie==1.1.1
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.2
omegaconf==2.3.0
openai==1.25.2
opencv-python==4.9.0.80
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
optuna==3.6.1
orjson==3.10.3
packaging==23.2
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
pooch==1.8.1
preshed==3.0.9
pretty-errors==1.2.25
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==24.4.0
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydash==7.0.7
Pygments==2.18.0
pyocr==0.8.5
pyparsing==3.1.2
pypdf==4.2.0
pytest==7.4.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rich==13.7.0
rpds-py==0.18.1
rstr==3.2.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.2.2
scikit-optimize==0.10.1
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
shutup==0.2.0
six==1.16.0
sklearn-pandas==2.2.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorpack==0.11
termcolor==2.4.0
texttable==1.7.0
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.5.3
tiktoken==0.5.2
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchinfo==1.8.0
torchmetrics==1.4.0
torchtext==0.18.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.2
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
wrapt==1.16.0
xlrd==2.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
